requirements.txt
certifi==2019.3.9
chardet==3.0.4
gitdb2==2.0.5
GitPython==2.1.11
idna==2.8
procfile-manager==0.0.2
python-dotenv==0.10.1
requests==2.21.0
servicefactory==1.0.2
smmap2==2.0.5
urllib3==1.24.1
Werkzeug==0.14.1
